Added more floating island types. Added Diving Helmet, Diving Gear, and Neptune's Shell.
Disabled Rocket Boots when in water.
Fixed running sound playing when not on solid blocks.

Removed outdated Localisation files (if you want to create add a translation just pull it).

MixinExtras has been added to make things easier to add, needed it for Neptune's Shell hiding air bubbles.

+++ src/main/java/terracraft/common/world/CorruptionPitCarver.java
@@ -34,7 +34,7 @@ public class CorruptionPitCarver extends WorldCarver<CorruptionPitCarverConfigur
         double h = corruptionPitCarverConfiguration.yScale.sample(random);
         float k = corruptionPitCarverConfiguration.shape.thickness.sample(random);
         int l = (int)((float)i * corruptionPitCarverConfiguration.shape.distanceFactor.sample(random));
-        this.doCarve(carvingContext, corruptionPitCarverConfiguration, chunkAccess, function, random.nextLong(), aquifer, d, (double)j, e, k, f, g, 0, l, h, carvingMask);
+        this.doCarve(carvingContext, corruptionPitCarverConfiguration, chunkAccess, function, random.nextLong(), aquifer, d, j, e, k, f, g, 0, l, h, carvingMask);
         return true;
     }
